Sort out ClientResponse and (de)serialization locations
