+++ serialization.h
@@ -67,6 +67,44 @@ inline void from_json(const nlohmann::json& j, Palette & palette)
 }
 
 //
+// ClientResponse serialization
+//
+
+inline void to_json(nlohmann::json &j, const ClientResponse &response)
+{
+    j ={
+            {"responseSize", response.size},
+            {"sequenceNumber", response.sequence},
+            {"flashVersion", response.flashVersion},
+            {"currentClock", response.currentClock},
+            {"oldestPacket", response.oldestPacket},
+            {"newestPacket", response.newestPacket},
+            {"brightness", response.brightness},
+            {"wifiSignal", response.wifiSignal},
+            {"bufferSize", response.bufferSize},
+            {"bufferPos", response.bufferPos},
+            {"fpsDrawing", response.fpsDrawing},
+            {"watts", response.watts}
+    };
+}
+
+inline void from_json(const nlohmann::json& j, ClientResponse& response) 
+{
+    response.size = j.at("responseSize").get<uint8_t>();
+    response.sequence = j.at("sequenceNumber").get<uint32_t>();
+    response.flashVersion = j.at("flashVersion").get<uint32_t>();
+    response.currentClock = j.at("currentClock").get<uint64_t>();
+    response.oldestPacket = j.at("oldestPacket").get<uint64_t>();
+    response.newestPacket = j.at("newestPacket").get<uint64_t>();
+    response.brightness = j.at("brightness").get<uint8_t>();
+    response.wifiSignal = j.at("wifiSignal").get<int8_t>();
+    response.bufferSize = j.at("bufferSize").get<uint32_t>();
+    response.bufferPos = j.at("bufferPos").get<uint32_t>();
+    response.fpsDrawing = j.at("fpsDrawing").get<float>();
+    response.watts = j.at("watts").get<float>();
+}
+
+//
 // Feature serialization
 //
